content/media/video/test/test_seek6.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: seek test 6</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id='v'
       src='seek.ogg'
       onloadedmetadata='return startTest();'
       onended='return playbackEnded();'
       onseeking='return seekStarted();'
       onseeked='return seekEnded();'></video>
<pre id="test">
<script class="testbody" type="text/javascript">
// Test for a seek, followed by a play before the seek completes, ensure we play at the end of the seek.
var startPassed = false;
var seekCount = 0;
function startTest() {
  var v = document.getElementById('v');
  v.currentTime=2;
}

function seekStarted() {
  startPassed = true;
  var v = document.getElementById('v');
  v.play();
  return false;
}

function seekEnded() {
  var v = document.getElementById('v');
  ok(v.currentTime>=2 && v.currentTime<=2.2, "First seek");
  return false;
}

function playbackEnded() {
  ok(true, "Playback ended");
  SimpleTest.finish();
  return false;
}

setTimeout(function () {
             ok(false, "Test timed out");
             SimpleTest.finish();
            }, 30000);

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
